requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
elevenlabs
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.51.0
googletrans
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
install==1.3.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
moviepy==1.0.3
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
noisereduce==3.0.2
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pedalboard==0.9.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytube==15.0.0
pyzmq==26.0.2
referencing==0.34.0
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.3
stack-data==0.6.3
threadpoolctl==3.4.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
